feat(vti-secrets): lift secret-store backends + integration onboarding into a shared crate (#501)
Extract the reusable SeedStore backends, the create_seed_store factory, and
SecretsConfig out of vta-service into a new shared `vti-secrets` crate so
external VTI integrations (e.g. vti-message-bridge) can onboard and store
secrets exactly like first-party services, without depending on the whole
vta-service binary crate.

Part 1 — extraction (pure, behaviour-preserving):
- New `vti-secrets` crate holds aws/gcp/azure/vault/k8s/keyring/config-seed/
  tee-kms/plaintext backends + create_seed_store + SecretsConfig, behind the
  same feature flags. Factory signature is now
  create_seed_store(&secrets, &data_dir) (no AppConfig dependency).
- vta-service re-exports them from keys::seed_store via a thin facade and keeps
  a create_seed_store(&AppConfig) wrapper, so every existing call site is
  unchanged. config.rs re-exports SecretsConfig. Backend files moved with
  git mv (history preserved); each vta-service feature now also activates the
  matching vti-secrets/<feature>.

Part 2 — onboarding helper (feature `onboarding`):
- vti_secrets::onboarding::IntegrationOnboarding wraps vta-sdk's SessionStore:
  ephemeral did:key mint -> pending-rotation -> connect()/connect_rest()
  auto-rotates off the temp DID on first auth. OnboardingTicket emits the
  exact `vta import-did --role application --context <ctx>` grant command.

Incidental: unify the duplicated `secure_file` owner-only file-hardening helper
into vti-common (vta-cli-common + vtc-service now re-export it), so vti-secrets'
plaintext backend doesn't pull in vta-cli-common.

Docs: root CLAUDE.md crate map/dep graph + docs/02-vta/secret-backends.md.
