+++ vti-secrets/src/seed_store/keyring.rs
@@ -1,8 +1,8 @@
 use std::future::Future;
 use std::pin::Pin;
 
-use crate::error::AppError;
 use tracing::debug;
+use vti_common::error::AppError;
 
 pub struct KeyringSeedStore {
     service: String,
